[22.06.27] DB data add: 데이터 베이스에 운동 데이터 이전꺼 추가, 어제 알고리즘 풀이한 부분까지 추가 진행

+++ 02.studyData/10.Algorithm/2022/백준코테/Algorithm/src/Algorithm/Algorithm/2020-카카오인턴십-수식최대화.cpp
@@ -47,8 +47,8 @@ long long solution(string expression) {
 			long long sum = 0;
 			for (int e = 0; e < calNumber.size(); e++) {
 				if (calNumber[e] == ordering[i][j]) {
-					int firstNum = stol (calNumber[e - 1]);
-					int secondNum = stol(calNumber[e + 1]);
+					long long firstNum = stoll (calNumber[e - 1]);
+					long long secondNum = stoll(calNumber[e + 1]);
 
 					if (calNumber[e] == "+") {
 						sum = firstNum + secondNum;
